refactor: Movie entity 수정 - Movies json file을 DB에 반영할 때 오류가 발생하여 score, popularity data type을 float에서 double로 변경

+++ community/src/main/java/com/movie/community/domain/Movie.java
@@ -33,7 +33,7 @@ public class Movie {
 	private String description;
 
 	@Column
-	private float score;
+	private double score;
 
 	@Column
 	private int voteCount;
@@ -45,7 +45,7 @@ public class Movie {
 	private LocalDate releasedDate;
 
 	@Column
-	private float popularity;
+	private double popularity;
 
 	@Column
 	private String genres;
